Renaming files, folders and classes
The old "guide" page is the new "wallet page"
