+++ lib/pages/wallet/guide_content.dart
