+++ lib/pages/wallet/widgets/barcode_scanner.dart
