+++ lib/pages/wallet/widgets/expandable_faq_item.dart
