+++ lib/pages/wallet/widgets/stacked_card_carousel.dart
